% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBShiny3.R
\name{NBShiny3}
\alias{NBShiny3}
\title{Launch 'NBShiny3' Interface}
\usage{
NBShiny3()
}
\value{
Nothing
}
\description{
NBShiny3() loads interactive user interface built using R 'shiny'.
}
\details{
Start NBShiny3

The interactive user interface is to provide an easy way for naive bayesian classification analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
NBShiny3()
}
}
\keyword{NBShiny3}
